added error message on date validation
